ReplaceOrderByExpressionWithEval
